requirements.txt
django==5.2.8
langchain-deepseek
python-dotenv
